Add functionality to UserLandingPage
After logging in, the user is taken to their main page where they can easily set an appointment.  They just have to select a day from the calendar which sorts contractors with availability that day. They can then select a contractor, service and time. Once all are selected a confirmation card appears with same functionality as on the Contractor page.