Fix delete user feedback issue
The action to delete feedback returned the deleted item to the dispatch rather than the ammended list.  I ammended the list and passed that to the store instead.  Then I changed the use effect to watch for feedback changes and also removed the loading from getFeedback so that wasn't popping up all the time.